Moved isClickOutside back to the top level utils directory as it is also used by the navbar
