+++ frontend/src/utils/Helpers/isClickOutside.js
